Remove app and routing module, tidy app
