+++ samples/msal-angular-v3-samples/angular-standalone-sample/README.md
@@ -10,7 +10,7 @@ This project was generated with [Angular CLI](https://github.com/angular/angular
 - Ensure [all pre-requisites](https://github.com/AzureAD/microsoft-authentication-library-for-js/blob/dev/lib/msal-angular/README.md) have been completed to run msal-angular.
 
 ### Configure the application
-- Open `./src/app/app.modules.ts` in an editor.
+- Open `./src/main.ts` in an editor.
 - Replace client id with the Application (client) ID from the portal registration, or use the currently configured lab registration. 
   - Optionally, you may replace any of the other parameters, or you can remove them and use the default values.
 
@@ -20,4 +20,4 @@ This project was generated with [Angular CLI](https://github.com/angular/angular
 - In the web page, click on the "Login" button. The app will automatically reload if you change any of the source files.
 
 ## Additional notes
-- The default interaction type for the sample is redirects. The sample can be configured to use redirects by changing the `interactionType` in `app.module.ts` to `InteractionType.Popup`. 
+- The default interaction type for the sample is redirects. The sample can be configured to use redirects by changing the `interactionType` in `main.ts` to `InteractionType.Popup`.
