--- samples/msal-angular-v3-samples/angular-standalone-sample/src/app/app-routing.module.ts
@@ -1,32 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-import { MsalGuard } from '@azure/msal-angular';
-import { BrowserUtils } from '@azure/msal-browser';
-import { ProfileComponent } from './profile/profile.component';
-import { HomeComponent } from './home/home.component';
-import { FailedComponent } from './failed/failed.component';
-
-const routes: Routes = [
-  {
-    path: 'profile',
-    component: ProfileComponent,
-    canActivate: [MsalGuard]
-  },
-  {
-    path: '',
-    component: HomeComponent
-  },
-  {
-    path: 'login-failed',
-    component: FailedComponent
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forRoot(routes, {
-    // Don't perform initial navigation in iframes or popups
-    initialNavigation: !BrowserUtils.isInIframe() && !BrowserUtils.isInPopup() ? 'enabledNonBlocking' : 'disabled' // Set to enabledBlocking to use Angular Universal
-  })],
-  exports: [RouterModule]
-})
-export class AppRoutingModule { }
