--- samples/msal-angular-v3-samples/angular-standalone-sample/src/app/failed/failed.component.html
@@ -1,3 +1,0 @@
-<p>
-    Login failed. Please try again.
-</p>
