+++ samples/msal-angular-v3-samples/angular-standalone-sample/src/app/failed/failed.component.ts
@@ -1,16 +1,9 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
-    selector: 'app-orders',
-    templateUrl: './failed.component.html',
-    styleUrls: ['./failed.component.css'],
+    selector: 'app-failed',
+    template: `<p>Login failed. Please try again.</p>`,
+    styleUrls: [],
     standalone: true
 })
-export class FailedComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
+export class FailedComponent {}
