+++ samples/msal-angular-v3-samples/angular-standalone-sample/src/index.html
@@ -11,7 +11,6 @@
 
 <body class="mat-typography">
   <app-root></app-root>
-  <app-redirect></app-redirect>
 </body>
 
 </html>
