+++ samples/msal-angular-v3-samples/angular-standalone-sample/src/main.ts
@@ -1,25 +1,88 @@
 import { enableProdMode, importProvidersFrom } from '@angular/core';
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
-
-import { MSALInstanceFactory, MSALInterceptorConfigFactory, MSALGuardConfigFactory, loggerCallback } from './app/app.module';
-import { environment } from './environments/environment';
-import { AppComponent } from './app/app.component';
+import { HTTP_INTERCEPTORS, withInterceptorsFromDi, provideHttpClient } from '@angular/common/http';
 import { MatMenuModule } from '@angular/material/menu';
 import { MatListModule } from '@angular/material/list';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatButtonModule } from '@angular/material/button';
-import { AppRoutingModule } from './app/app-routing.module';
 import { provideNoopAnimations } from '@angular/platform-browser/animations';
 import { BrowserModule, bootstrapApplication } from '@angular/platform-browser';
-import { environment as environment_1 } from 'src/environments/environment';
-import { IPublicClientApplication, PublicClientApplication, InteractionType, BrowserCacheLocation, LogLevel } from '@azure/msal-browser';
-import { MsalInterceptor, MSAL_INSTANCE, MsalInterceptorConfiguration, MsalGuardConfiguration, MSAL_GUARD_CONFIG, MSAL_INTERCEPTOR_CONFIG, MsalService, MsalGuard, MsalBroadcastService, MsalModule } from '@azure/msal-angular';
-import { HTTP_INTERCEPTORS, withInterceptorsFromDi, provideHttpClient } from '@angular/common/http';
+import { Route, provideRouter, withDisabledInitialNavigation, withEnabledBlockingInitialNavigation } from '@angular/router';
+import { MsalInterceptor, MSAL_INSTANCE, MsalInterceptorConfiguration, MsalGuardConfiguration, MSAL_GUARD_CONFIG, MSAL_INTERCEPTOR_CONFIG, MsalService, MsalGuard, MsalBroadcastService } from '@azure/msal-angular';
+import { IPublicClientApplication, PublicClientApplication, InteractionType, BrowserCacheLocation, LogLevel, BrowserUtils } from '@azure/msal-browser';
+import { AppComponent } from './app/app.component';
+import { HomeComponent } from './app/home/home.component';
+import { ProfileComponent } from './app/profile/profile.component';
+import { FailedComponent } from './app/failed/failed.component';
+import { environment } from './environments/environment';
 
-const protectedResourceMap = new Map<string, Array<string>>();
 const isIE = window.navigator.userAgent.indexOf("MSIE ") > -1 || window.navigator.userAgent.indexOf("Trident/") > -1;
 
+export function loggerCallback(logLevel: LogLevel, message: string) {
+    console.log(message);
+}
 
+export function MSALInstanceFactory(): IPublicClientApplication {
+    return new PublicClientApplication({
+      auth: {
+        clientId: environment.msalConfig.auth.clientId,
+        authority: environment.msalConfig.auth.authority,
+        redirectUri: '/',
+        postLogoutRedirectUri: '/'
+      },
+      cache: {
+        cacheLocation: BrowserCacheLocation.LocalStorage,
+        storeAuthStateInCookie: isIE, // set to true for IE 11. Remove this line to use Angular Universal
+      },
+      system: {
+        allowNativeBroker: false, // Disables WAM Broker
+        loggerOptions: {
+          loggerCallback,
+          logLevel: LogLevel.Info,
+          piiLoggingEnabled: false
+        }
+      }
+    });
+}
+
+export function MSALInterceptorConfigFactory(): MsalInterceptorConfiguration {
+    const protectedResourceMap = new Map<string, Array<string>>();
+    protectedResourceMap.set(environment.apiConfig.uri, environment.apiConfig.scopes);
+  
+    return {
+      interactionType: InteractionType.Redirect,
+      protectedResourceMap
+    };
+  }
+  
+  export function MSALGuardConfigFactory(): MsalGuardConfiguration {
+    return { 
+      interactionType: InteractionType.Redirect,
+      authRequest: {
+        scopes: [...environment.apiConfig.scopes]
+      },
+      loginFailedRoute: '/login-failed'
+    };
+}
+
+const initialNavigation = !BrowserUtils.isInIframe() && !BrowserUtils.isInPopup() 
+    ? withEnabledBlockingInitialNavigation() // Set to enabledBlocking to use Angular Universal
+    : withDisabledInitialNavigation(); 
+
+export const Routes: Route[] = [
+    {
+        path: 'profile',
+        component: ProfileComponent,
+        canActivate: [MsalGuard]
+    },
+    {
+        path: '',
+        component: HomeComponent
+    },
+    {
+        path: 'login-failed',
+        component: FailedComponent
+    }
+    ];
 
 if (environment.production) {
   enableProdMode();
@@ -27,7 +90,10 @@ if (environment.production) {
 
 bootstrapApplication(AppComponent, {
     providers: [
-        importProvidersFrom(BrowserModule, AppRoutingModule, MatButtonModule, MatToolbarModule, MatListModule, MatMenuModule, MsalModule),
+        importProvidersFrom(BrowserModule, MatButtonModule, MatToolbarModule, MatListModule, MatMenuModule),
+        provideRouter(Routes, initialNavigation),
+        provideNoopAnimations(),
+        provideHttpClient(withInterceptorsFromDi()),
         {
             provide: HTTP_INTERCEPTORS,
             useClass: MsalInterceptor,
@@ -47,9 +113,7 @@ bootstrapApplication(AppComponent, {
         },
         MsalService,
         MsalGuard,
-        MsalBroadcastService,
-        provideNoopAnimations(),
-        provideHttpClient(withInterceptorsFromDi())
+        MsalBroadcastService
     ]
 })
   .catch(err => console.error(err));
